move to lookup.h
